+++ src/base/clptur.f90
@@ -2050,7 +2050,7 @@ double precision tetmax, tetmin, tplumx, tplumn
 integer          ivar, f_id, b_f_id, isvhbl
 integer          f_id_ut, f_id_al
 integer          ifac, iel, isou, jsou
-integer          ifcvsl, itplus, itstar
+integer          iscacp, ifcvsl, itplus, itstar
 
 double precision cpp, rkl, prdtl, visclc, romc, tplus, cofimp, cpscv
 double precision distfi, distbf, fikis, hint_al, heq, hflui, hext
@@ -2242,8 +2242,11 @@ else
   endif
 endif
 
+! Does the scalar behave as a temperature ?
+call field_get_key_int(f_id, kscacp, iscacp)
+
 ! retrieve turbulent Schmidt value for current scalar
-call field_get_key_double(ivarfl(isca(iscal)), ksigmas, turb_schmidt)
+call field_get_key_double(f_id, ksigmas, turb_schmidt)
 
 ! --- Loop on boundary faces
 do ifac = 1, nfabor
@@ -2269,7 +2272,7 @@ do ifac = 1, nfabor
     distbf = distb(ifac)
 
     cpp = 1.d0
-    if (iscacp(iscal).eq.1) then
+    if (iscacp.eq.1) then
       if (icp.ge.0) then
         cpp = cpro_cp(iel)
       else
@@ -2440,7 +2443,7 @@ do ifac = 1, nfabor
     distbf = distb(ifac)
 
     cpp = 1.d0
-    if (iscacp(iscal).eq.1) then
+    if (iscacp.eq.1) then
       if (icp.ge.0) then
         cpp = cpro_cp(iel)
       else
@@ -2535,7 +2538,7 @@ do ifac = 1, nfabor
           endif
 
         ! Temperature
-        elseif (iscacp(iscal).eq.1) then
+        elseif (iscacp.eq.1) then
           exchange_coef = hflui
         endif
       endif
@@ -2781,7 +2784,7 @@ double precision, dimension(:) :: byplus, bdplus, buk
 
 integer          ivar, f_id, isvhbl
 integer          ifac, iel
-integer          ifcvsl
+integer          iscacp, ifcvsl
 
 double precision cpp, rkl, prdtl, visclc, romc, cofimp
 double precision distbf, heq, yptp, hflui, hext
@@ -2831,8 +2834,11 @@ if (icp.ge.0) then
   call field_get_val_s(icp, cpro_cp)
 endif
 
+! Does the vector behave as a temperature ?
+call field_get_key_int(f_id, kscacp, iscacp)
+
 ! retrieve turbulent Schmidt value for current vector
-call field_get_key_double(ivarfl(isca(iscal)), ksigmas, turb_schmidt)
+call field_get_key_double(f_id, ksigmas, turb_schmidt)
 
 isvhbl = 0
 if (iscal.eq.isvhb) then
@@ -2871,7 +2877,7 @@ do ifac = 1, nfabor
     xnuii = visclc / romc
 
     cpp = 1.d0
-    if (iscacp(iscal).eq.1) then
+    if (iscacp.eq.1) then
       if (icp.ge.0) then
         cpp = cpro_cp(iel)
       else